+++ src/Core/Models/Strings.cs
@@ -572,6 +572,8 @@ namespace AccessibleArena.Core.Models
         public static string CardInfoPowerToughness => L.Get("CardInfoPowerToughness");
         public static string CardInfoType => L.Get("CardInfoType");
         public static string CardInfoRules => L.Get("CardInfoRules");
+        public static string LoyaltyPlus => L.Get("LoyaltyPlus");
+        public static string LoyaltyMinus => L.Get("LoyaltyMinus");
         public static string CardInfoFlavor => L.Get("CardInfoFlavor");
         public static string CardInfoRarity => L.Get("CardInfoRarity");
         public static string CardInfoArtist => L.Get("CardInfoArtist");